Add files via upload
This is a data analysis of a renting bikes business with data provided by a real company located in Chicago. Data is open and available here: 
https://divvy-tripdata.s3.amazonaws.com/index.html
I analized the patterns differences between casual users and annual members in order to find insights that could help the marketing department to make a strategy aimed to convert casual riders into annual members.